fix circular dependency on topic-add
